Conflicts:
	src/main/java/com/formation/controllers/OrderController.java

+++ src/main/java/com/formation/controllers/PictureController.java
@@ -14,7 +14,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.formation.dto.picture.PictureFull;
+import com.formation.exceptions.NotAuthorizedException;
 import com.formation.persistence.entities.Picture;
+import com.formation.services.IAuthChecker;
 import com.formation.services.IPictureService;
 
 @RestController
@@ -27,8 +29,12 @@ public class PictureController {
 	@Autowired
 	IPictureService service;
 	
+	@Autowired
+	private IAuthChecker authchecker; 
+	
 	@GetMapping 
 	public List<PictureFull> findAll() {
+		if (authchecker.getCurrentAdmin() == null) throw new NotAuthorizedException("Vous n'avez pas les droits pour cette action");
 		return service.findAll()
 				.stream()
 				.map(pi -> mapper.map(pi, PictureFull.class))
@@ -37,16 +43,20 @@ public class PictureController {
 	
 	@GetMapping (path = "/{id}")
 	public PictureFull findOne(@PathVariable(name = "id") Long id) {
+		if (authchecker.getCurrentAdmin() == null) throw new NotAuthorizedException("Vous n'avez pas les droits pour cette action");
 		return mapper.map(service.findOne(id), PictureFull.class);		
 	}
 	
 	@DeleteMapping(path = "/{id}")
 	public boolean deleteById(@PathVariable(name = "id") Long id) {
+		if (authchecker.getCurrentAdmin() == null) throw new NotAuthorizedException("Vous n'avez pas les droits pour cette action");
 		return service.deleteById(id);
 	}
 	
 	@PostMapping
 	public PictureFull save(@RequestBody PictureFull pictureF) {
+		if (authchecker.getCurrentAdmin() == null) throw new NotAuthorizedException("Vous n'avez pas les droits pour cette action");
+		
 		Picture picture = service.save(mapper.map(pictureF, Picture.class));
 		PictureFull pictureFSauve = mapper.map(picture, PictureFull.class);
 		return pictureFSauve;
